health-and-life-sciences-ai-suite/microservices/aggregator-service/requirements.txt

grpcio
grpcio-tools
protobuf
